refactor: re-typings, because why not
