+++ src/main.ts
@@ -1,2 +1,3 @@
 export * from './errors'
 export * from './mediawiki'
+export * from './types'
