+++ src/mediawiki/_base/Bot.ts
@@ -1,17 +1,9 @@
-import type {
-	MediaWikiRequest,
-	MediaWikiResponse,
-	ReducedRequest
-} from '../../types'
-import {
-	ErrorManager
-} from '../../errors'
+import type { APIError, BlockRequest, BlockResponse, DeleteRequest, DeleteResponse, EditRequest, EditResponse, LoginRequest, LoginResponse, MoveRequest, MoveResponse, NoActionToken, NoJSONRequest, ProtectRequest, ProtectResponse, UploadRequest, UploadResponse } from '../../types'
+import { ErrorManager } from '../../errors'
 import fetch from 'node-fetch'
 import fs from 'fs-extra'
 import tmp from 'tmp-promise'
-import type {
-	Wiki
-} from './Wiki'
+import type { Wiki } from './Wiki'
 
 export class Bot<WikiType extends Wiki = Wiki> {
 	readonly #password: string
@@ -35,19 +27,19 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		this.#wiki = wiki
 	}
 
-	public async block( params: ReducedRequest<MediaWikiRequest.Block> ): Promise<MediaWikiResponse.Block> {
+	public async block( params: NoActionToken<BlockRequest> ): Promise<BlockResponse> {
 		const token = await this.getCSRFToken()
-		return this.wiki.post<MediaWikiResponse.Block, MediaWikiRequest.Block>( {
+		return this.wiki.post<BlockResponse, NoJSONRequest<BlockRequest>>( {
 			...params,
 			action: 'block',
 			token
 		} )
 	}
 
-	public async delete( params: ReducedRequest<MediaWikiRequest.Delete> ): Promise<MediaWikiResponse.Delete> {
+	public async delete( params: NoActionToken<DeleteRequest> ): Promise<DeleteResponse> {
 		const token = await this.getCSRFToken()
 
-		const req = await this.#wiki.post<MediaWikiResponse.Delete | MediaWikiResponse.ApiError>( {
+		const req = await this.#wiki.post<DeleteResponse | APIError>( {
 			...params,
 			action: 'delete',
 			token
@@ -61,10 +53,10 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		return req
 	}
 
-	public async edit( params: ReducedRequest<MediaWikiRequest.Edit> ): Promise<MediaWikiResponse.Edit> {
+	public async edit( params: NoActionToken<EditRequest> ): Promise<EditResponse> {
 		const token = await this.getCSRFToken()
 
-		const req = await this.#wiki.post<MediaWikiResponse.Edit | MediaWikiResponse.ApiError>( {
+		const req = await this.#wiki.post<EditResponse | APIError>( {
 			...params,
 			action: 'edit',
 			assert: params.bot ? 'bot' : 'user',
@@ -102,7 +94,7 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		const tokenreq = await this.#wiki.getToken( 'login' )
 		const lgtoken = tokenreq.query.tokens.logintoken
 
-		const res = await this.#wiki.post<MediaWikiResponse.Login, MediaWikiRequest.Login>( {
+		const res = await this.#wiki.post<LoginResponse, LoginRequest>( {
 			action: 'login',
 			lgname: this.#username,
 			lgpassword: this.#password,
@@ -123,9 +115,9 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		this.#wiki.request.clear( this.#wiki.api )
 	}
 
-	public async move( params: ReducedRequest<MediaWikiRequest.Move> ): Promise<MediaWikiResponse.Move> {
+	public async move( params: NoActionToken<MoveRequest> ): Promise<MoveResponse> {
 		const token = await this.getCSRFToken()
-		const req = await this.#wiki.post<MediaWikiResponse.Move | MediaWikiResponse.ApiError>( {
+		const req = await this.#wiki.post<MoveResponse | APIError>( {
 			...params,
 			action: 'move',
 			token
@@ -139,7 +131,7 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		return req
 	}
 
-	public async protect( params: ReducedRequest<MediaWikiRequest.Protect> ): Promise<MediaWikiResponse.Protect> {
+	public async protect( params: NoActionToken<ProtectRequest> ): Promise<ProtectResponse> {
 		const token = await this.getCSRFToken()
 
 		return this.wiki.post( {
@@ -157,7 +149,7 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		return this.wiki.purge( titles )
 	}
 
-	public async unblock( params: ReducedRequest<MediaWikiRequest.Block> ): Promise<MediaWikiResponse.Block> {
+	public async unblock( params: NoActionToken<BlockRequest> ): Promise<BlockResponse> {
 		const token = await this.getCSRFToken()
 		return this.wiki.post( {
 			...params,
@@ -166,10 +158,10 @@ export class Bot<WikiType extends Wiki = Wiki> {
 		} )
 	}
 
-	public async upload( params: ReducedRequest<MediaWikiRequest.Upload> ): Promise<MediaWikiResponse.Upload> {
+	public async upload( params: NoActionToken<UploadRequest> ): Promise<UploadResponse> {
 		const token = await this.getCSRFToken()
 
-		const req = await this.#wiki.post<MediaWikiResponse.Upload | MediaWikiResponse.ApiError>( {
+		const req = await this.#wiki.post<UploadResponse | APIError>( {
 			...params,
 			action: 'upload',
 			token
@@ -192,7 +184,7 @@ export class Bot<WikiType extends Wiki = Wiki> {
 	 */
 	public async uploadFromUrl( {
 		filename, url
-	}: { filename: string, url: string } ): Promise<MediaWikiResponse.Upload | undefined> {
+	}: { filename: string, url: string } ): Promise<UploadResponse | undefined> {
 		const image = await fetch( url )
 		if ( !image.ok ) return undefined
