--- src/types/index.ts
@@ -1,8 +1,0 @@
-export * from './MetaRequest'
-export * from './MetaResponse'
-export * from './QueryItem'
-export * from './QueryRequest'
-export * from './QueryResponse'
-export * from './Request'
-export * from './Response'
-export * from './utils'
