+++ tsconfig.json
@@ -1,11 +1,13 @@
 {
 	"exclude": [ "dist", "node_modules" ],
-	"include": [ "src" ],
+	"include": [ "src", "types.d.ts" ],
 	"compilerOptions": {
 		"module": "CommonJS",
 		"target": "ES6",
 		"skipLibCheck": true,
 		"strict": true,
+		"declaration": true,
+		"sourceMap": true,
 
 		/* Paths */
 		"outDir": "./dist",
